+++ tests/util/test_stringUtils.py
@@ -1,5 +1,5 @@
 import unittest
-from yacg.util.stringUtils import toUpperCamelCase
+from yacg.util.stringUtils import toUpperCamelCase, toName
 
 
 class TestStringUtils (unittest.TestCase):
@@ -10,3 +10,19 @@ class TestStringUtils (unittest.TestCase):
     def testUpperCamelCase2(self):
         t = toUpperCamelCase('IamAnotherTest')
         self.assertEqual('IamAnotherTest', t)
+
+    def testToName(self):
+        t = toName('IamAnotherTest')
+        self.assertEqual('IamAnotherTest', t)
+        t = toName('1')
+        self.assertEqual('_1', t)
+        t = toName('1.1')
+        self.assertEqual('_1_1', t)
+        t = toName('1/1/')
+        self.assertEqual('_1_1_', t)
+        t = toName('1:1:')
+        self.assertEqual('_1_1_', t)
+        t = toName('a:1')
+        self.assertNotEqual('a_1_', t)
+        t = toName('a:1')
+        self.assertEqual('a_1', t)